refactor(renderer): platform定数をinfrastructure/adaptersに移動し、既存のアダプターに統合
