+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tag-util",
-  "version": "1.1.1",
+  "version": "1.1.2",
   "description": "Functional FLAC metadata editor with Svelte 5 TagState",
   "type": "module",
   "main": "./out/main/index.js",
