+++ src/renderer/src/components/StatusBar.svelte
@@ -3,7 +3,7 @@
   import { ChevronUp, X, Check, TriangleAlert, List, type LucideProps } from '@lucide/svelte';
   import { tooltip } from '@renderer/actions/tooltip';
   import { UI_TOKENS } from '@renderer/constants/design-system';
-  import { IS_MAC } from '@renderer/constants/platform';
+  import { IS_MAC } from '@renderer/infrastructure/adapters/platform-adapter';
   import { logStore } from '@renderer/stores/log-store.svelte';
   import { KeyboardHandler } from '@renderer/utils/keyboard-handler';
   import { formatTimeWithMs } from '@shared/utils/date';
