+++ src/renderer/src/infrastructure/adapters/platform-adapter.ts
@@ -1,9 +1,17 @@
-import { type Platform } from '@shared/platform';
+let isMacInternal = false;
 
 /**
- * 実行環境（プラットフォーム）を取得します。
- * Electron の IPC 通信を使用して情報を取得します。
+ * アプリケーションが macOS 上で動作しているかどうか。
+ * 初期化前は false です。初期化には initializePlatform() を呼び出す必要があります。
+ * 外部からは読み取り専用です。
  */
-export const getPlatform = async (): Promise<Platform> => {
-  return await window.api.getPlatform();
+export { isMacInternal as IS_MAC };
+
+/**
+ * プラットフォーム情報を初期化します。
+ * 外部環境から情報を取得するため、非同期で実行されます。
+ */
+export const initializePlatform = async (): Promise<void> => {
+  const platform = await window.api.getPlatform();
+  isMacInternal = platform.isMac;
 };
