+++ src/renderer/src/main.ts
@@ -3,7 +3,7 @@ import './index.css';
 
 import App from './App.svelte';
 import { initializeGlobalErrorHandler } from './infrastructure/error-handler';
-import { initializePlatform } from './constants/platform';
+import { initializePlatform } from './infrastructure/adapters/platform-adapter';
 
 const init = async (): Promise<void> => {
   initializeGlobalErrorHandler();
